Jambase: Change default ARFLAGS and enable NOARSCAN by default.
On most systems, "D" (deterministic) is now the default archiving mode,
and so specifying "u" will throw a warning that "D" is being used anyway.
Jam previously assumed, as it was passing "u", that the timestamps inside
.a's were valid when they were actually 0, which was leading to spurious
rebuilds of .as.

After this change, we assume archives cannot be scanned for timestamps,
which should stop the spurious rebuilds of targets on most systems.