jam: mirror Jamfile in Makefile.
Without these changes, `jam0` is different to `bin.*/jam`, a subtle
difference that was triggering errors in CI due to use of `jam0`.
In particular, the omission of `OPT_JAM_TARGETS_VARIABLE_EXT`
results in a `jam0` that cannot build Haiku, as `$(JAM_TARGETS)`
will not be defined, and is required by Haiku's build system.